chore(app-shell): extract logic into util
