+++ packages/application-shell/src/utils/get-previous-project-key/get-previous-project-key.js
@@ -1,0 +1,12 @@
+import selectProjectKeyFromLocalStorage from '../select-project-key-from-local-storage';
+
+const getPreviousProjectKey = defaultProjectKeyOfUser => {
+  const previouslyUsedProjectKeyFromLocalStorage = selectProjectKeyFromLocalStorage();
+  if (previouslyUsedProjectKeyFromLocalStorage)
+    return previouslyUsedProjectKeyFromLocalStorage;
+  if (defaultProjectKeyOfUser) return defaultProjectKeyOfUser;
+
+  return null;
+};
+
+export default getPreviousProjectKey;
