+++ packages/application-shell/src/utils/get-previous-project-key/index.js
@@ -1,0 +1,1 @@
+export { default } from './get-previous-project-key';
